Move implementation of cancellation_token into submodule file
